Move assert_equal into expect
